requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==12.1.0
binaryornot==0.4.4
bitarray==2.9.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
Cython==3.0.10
decorator==5.1.1
dnspython==2.6.1
edge-tts==6.1.12
email_validator==2.2.0
exceptiongroup==1.2.1
fairseq==0.12.2
faiss-cpu==1.7.3
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg-python==0.2.0
filelock==3.15.4
flet==0.23.1
flet-core==0.23.1
flet-runtime==0.23.1
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hydra-core==1.0.7
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.0.6
orjson==3.10.5
packaging==23.2
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.0
praat-parselmouth==0.4.3
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pyworld==0.3.4
PyYAML==6.0.1
qrcode==7.4.2
regex==2024.5.15
repath==0.9.0
requests==2.31.0
resampy==0.4.3
rich==13.7.1
rsa==4.9
rvc-python==0.0.7
sacrebleu==2.4.2
scikit-learn==1.5.0
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchcrepe==0.0.23
tqdm==4.66.4
triton==2.3.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
